refactor: classify clipboard helper, query
